Move BackendRequestException into infrastructure
