+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/exceptions/CouldNotArchiveException.kt
@@ -1,4 +1,4 @@
 package fr.gouv.cacem.monitorenv.domain.exceptions
 
-@Deprecated("Use `BackendUsageException` instead.")
+@Deprecated("Use `BackendException` instead.")
 class CouldNotArchiveException(message: String) : RuntimeException(message)
