+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/exceptions/CouldNotDeleteException.kt
@@ -1,4 +1,4 @@
 package fr.gouv.cacem.monitorenv.domain.exceptions
 
-@Deprecated("Use `BackendUsageException` instead.")
+@Deprecated("Use `BackendException` instead.")
 class CouldNotDeleteException(message: String) : RuntimeException(message)
