+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/exceptions/EntityConversionException.kt
@@ -1,5 +1,5 @@
 package fr.gouv.cacem.monitorenv.domain.exceptions
 
-@Deprecated("Use `BackendInternalException` or `BackendRequestException` instead (depending on the case).")
+@Deprecated("Use `BackendException` with the right code instead (depending on the case).")
 class EntityConversionException(message: String, cause: Throwable? = null) :
     Throwable(message, cause)
