+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/exceptions/NotFoundException.kt
@@ -1,5 +1,5 @@
 package fr.gouv.cacem.monitorenv.domain.exceptions
 
-@Deprecated("Use `BackendUsageException` instead.")
+@Deprecated("Use `BackendException` instead.")
 class NotFoundException(message: String, cause: Throwable? = null) :
     Throwable(message, cause)
