+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/exceptions/ReportingAlreadyAttachedException.kt
@@ -1,5 +1,5 @@
 package fr.gouv.cacem.monitorenv.domain.exceptions
 
 /** Thrown when attempting to attach a reporting already attached to a mission. */
-@Deprecated("Use `BackendUsageException` instead.")
+@Deprecated("Use `BackendException` instead.")
 class ReportingAlreadyAttachedException(message: String) : RuntimeException(message)
