+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/utils/validateId.kt
@@ -1,8 +1,8 @@
 package fr.gouv.cacem.monitorenv.infrastructure.api.endpoints.utils
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import fr.gouv.cacem.monitorenv.domain.exceptions.BackendRequestErrorCode
-import fr.gouv.cacem.monitorenv.domain.exceptions.BackendRequestException
+import fr.gouv.cacem.monitorenv.infrastructure.exceptions.BackendRequestErrorCode
+import fr.gouv.cacem.monitorenv.infrastructure.exceptions.BackendRequestException
 
 fun validateId(requestDataAsJson: String, idPropName: String, idFromRequestPath: Int, objectMapper: ObjectMapper) {
     val requestDataAsJsonNode = objectMapper.readTree(requestDataAsJson)
@@ -32,7 +32,7 @@ fun validateId(requestDataAsJson: String, idPropName: String, idFromRequestPath:
     val requestDataId = idAsJsonNode.asInt()
     if (requestDataId != idFromRequestPath) {
         throw BackendRequestException(
-            BackendRequestErrorCode.WRONG_REQUEST_BODY_PROPERTY_TYPE,
+            BackendRequestErrorCode.BODY_ID_MISMATCH_REQUEST_PATH_ID,
             "Request data `$idPropName` ('$requestDataId') doesn't match the {$idPropName} in the request path ('$idFromRequestPath').",
         )
     }
