--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/exceptions/InfrastructureErrorCode.kt
@@ -1,3 +1,0 @@
-package fr.gouv.cacem.monitorenv.infrastructure.database.repositories.exceptions
-
-enum class InfrastructureErrorCode
